Create Kaltura Retrieve Entry Data by Reference ID
Reads in a parameters file and a file containing a list of Kaltura Reference IDs, and writes out a file with details about the Kaltura video file or files retrieved with each Reference ID. Requests.POST used following examples in Kaltura API documentation. The parameters file must contain a valid Kaltura Partner ID, Administrative Secret, and Administrative User ID in order to gain access to the data. The program uses these parameters to retrieve a session token that is used to validate access in all subsequent web requests.