Create Zoom List All Users
Python code using Zoom API to retrieve a list of all users in an instance of Zoom. The program reads a parameters file that must contain a JSON Web Token (JWT) generated on Marketplace.zoom.us. The program follows examples from Zoom API documentation that use http.client in Python. The list of users is likely long, so the program uses the NEXT_PAGE_TOKEN received in one API Call to get another part of the list in the next API call.