Create Zoom List Recordings for One User
Python code using Zoom API to retrieve a list of one user’s recordings within a range of dates. The program reads a parameters file that must contain a JSON Web Token (JWT) generated on Marketplace.zoom.us. The parameters file also specifies a user ID (email address) and start and stop date. A maximum of one month or 30 days can be retrieved in one execution. This code does not handle multiple pages of data returned from Zoom, but displays a message and ends if an additional page was available.