Create Zoom List Recordings for Multiple Users
Python code using Zoom API to retrieve a list of  recordings belonging to multiple users and any range of dates. The program reads in a parameters file that specifies a JSON web token created on Marketplace.zoom.us, and a range of dates of interest. It reads in another file that includes the user IDs (email addresses) of the desired users. The program retrieves each user’s list of recordings in one-month chunks, with the NEXT_PAGE_TOKEN used to retrieve multiple pages of web response when needed.