Create Kaltura List Base Entries
Python program to list Kaltura Base Entries using the Kaltura API. The Kaltura API limits the number of entries returnable by paging through web responses to 10,000. For this reason, the program retrieves entries by date created. If there are more than 10,000 entries created within a 24-hour period, the program will end with an error message. The solution would be to adjust the code to retrieve a smaller time period, such as 12 hours, by any one query.  The Kaltura API seems only to return entries that are marked with a status of “Ready,” unless the query explicitly requests other statuses. The program makes one call to the API to establish a session token for subsequent queries. The program then makes a query to find the earliest creation date of base entries, then a third query to find the most recent creation date together with a preliminary count of entries. The program then loops through date ranges making one query to obtain the number of entries created within the range, then as many additional queries as needed to page through all those entries.