Create Zoom list all videos by date
The program lists all the Zoom video recording files within a specified date range.
Because the Zoom API limits any one query to a single user and a maximum date range of 30 days, the program must loop through multiple data ranges and a list of all users to make multiple queries. The Python Requests library and Zoom API are used.